docs: Distinguish application name
Distinguish between application name and executable name for the Vulkan
Info and Vulkan Cube applications.

Use Vulkan Info and Vulkan Cube when referring to the application and
vulkaninfo and vkcube/vkcubepp when referring to the command,
respectively.

This change was provoked by a desired to refer to applications in
consistent manner throughout the Vulkan SDK documentation.